backend/requirements.txt
fastapi
uvicorn
python-dotenv
youtube-transcript-api
langchain
langchain-community
langchain-groq
langchain-core
langchain-text-splitters
faiss-cpu
fastembed
pydantic